requirements.txt
# Core dependencies
streamlit==1.28.1
pandas==2.1.3
numpy==1.24.3
plotly==5.17.0

# Twitter/Social Media
tweepy==4.14.0
requests==2.31.0
beautifulsoup4==4.12.2

# AI/ML Libraries
openai==1.3.7
anthropic==0.7.7
sentence-transformers==2.2.2
transformers==4.35.2
torch==2.1.1
scikit-learn==1.3.2
vaderSentiment==3.3.2
spacy==3.7.2

# Data Processing
python-dotenv==1.0.0
sqlite3
json5==0.9.14
dateparser==1.2.0
pytz==2023.3

# Web scraping & APIs
selenium==4.15.2
webdriver-manager==4.0.1

# Alerts & Notifications
slack-sdk==3.23.0
discord-webhook==1.3.0

# Utilities
schedule==1.2.0
asyncio
aiohttp==3.9.1
logging
typing-extensions==4.8.0
